clean ups on stories
