+++ src/js/components/Grid/stories/NColumn.js
@@ -1,0 +1,26 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Grommet, Box, Grid } from 'grommet';
+import { grommet } from 'grommet/themes';
+
+const NColumnGrid = () => (
+  <Grommet theme={grommet} full>
+    <Grid
+      columns={{
+        count: 6,
+        size: 'auto',
+      }}
+      gap="small"
+    >
+      <Box background="brand">Item 1</Box>
+      <Box background="brand">Item 2</Box>
+      <Box background="brand">Item 3</Box>
+      <Box background="brand">Item 4</Box>
+      <Box background="brand">Item 5</Box>
+      <Box background="brand">Item 6</Box>
+    </Grid>
+  </Grommet>
+);
+
+storiesOf('Grid', module).add('N-column layout', () => <NColumnGrid />);
